비밀번호 숨김 기능 파일 분리
