+++ app/src/main/java/org/sopt/at/ui/theme/ShowHidePasswordTextField.kt
@@ -1,0 +1,80 @@
+package org.sopt.at.ui.theme
+
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Visibility
+import androidx.compose.material.icons.filled.VisibilityOff
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextField
+import androidx.compose.material3.TextFieldDefaults
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.input.KeyboardType
+import androidx.compose.ui.text.input.PasswordVisualTransformation
+import androidx.compose.ui.text.input.VisualTransformation
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun ShowHidePasswordTextField(){
+    var password by remember { mutableStateOf("") }
+    var showPassword by remember { mutableStateOf(false) }
+
+    TextField(
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(horizontal = 20.dp)
+            .height(55.dp),
+        value = password,
+        onValueChange = {password = it},
+        placeholder = {
+            Text(text = "비밀번호")
+        },
+        singleLine = true,
+        shape = RoundedCornerShape(5.dp),
+        colors = TextFieldDefaults.colors(
+            focusedTextColor = GrayHintText,
+            unfocusedTextColor = GrayHintText,
+            focusedPlaceholderColor = GrayHintText,
+            unfocusedPlaceholderColor = GrayHintText,
+            focusedContainerColor = GrayEdit,
+            unfocusedContainerColor = GrayEdit,
+            // 하단 밑줄 사라지게
+            focusedIndicatorColor = Color.Transparent,
+            unfocusedIndicatorColor = Color.Transparent
+        ),
+        visualTransformation = if (showPassword){
+            VisualTransformation.None
+        } else{
+            PasswordVisualTransformation()
+        },
+        keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Password),
+        trailingIcon = {
+            if (showPassword){
+                IconButton(onClick = {showPassword = false}) {
+                    Icon(
+                        imageVector = Icons.Filled.Visibility,
+                        contentDescription = "hide_password"
+                    )
+                }
+            } else {
+                IconButton(onClick = {showPassword = true}) {
+                    Icon(
+                        imageVector = Icons.Filled.VisibilityOff,
+                        contentDescription = "hide_password"
+                    )
+                }
+            }
+        }
+    )
+}
